// src/interfaces/firebase/IPeople.ts
export interface IPeople {
    id?: string;
    name: string;
    age: number;
    contact: number;
    ministry: string;
    gender: string;
    birthday: Date | "";
  }

  export interface IPeoplePublic {
    id?: string;
    name: string;
    age: number;
    contact: number;
    ministry: string;
    gender: string;
    birthday: Date | "";
  }